+++ src/de_datalake_bulletin_dataload/definitions.py
@@ -1,13 +1,21 @@
 from dagster import Definitions, EnvVar
-from de_datalake_bulletin_dataload.defs.resources import HTTPClientResource, ParquetExportResource, AWSS3Resource
-from de_datalake_bulletin_dataload.defs.assets import fetch_export_pages_data
+from de_datalake_bulletin_dataload.defs.resources import HTTPClientResource, ParquetExportResource, AWSS3Resource, ConfigResource
+from de_datalake_bulletin_dataload.defs.assets import fetch_export_pages_data, fetch_export_media_data
+from de_datalake_bulletin_dataload.defs.sensors import bulletin_data_sensor
 
 defs = Definitions(
-    assets=[fetch_export_pages_data],
+    assets=[fetch_export_pages_data, fetch_export_media_data],
+    sensors=[bulletin_data_sensor],
     resources={
+        "get_config": ConfigResource(
+            config_path=EnvVar("FETCH_CONFIG_PATH")
+        ),
         "http_client": HTTPClientResource(
             user_agent=EnvVar("USER_AGENT"),
-            base_url=EnvVar("BULLETIN_PAGES_WP_BASE_URL")
+            base_url=EnvVar("BULLETIN_WP_BASE_URL"),
+            get_config=ConfigResource(
+                config_path=EnvVar("FETCH_CONFIG_PATH")
+            )
         ),
         "parquet_export_path": ParquetExportResource(
             export_folder_path=EnvVar("PARQUET_EXPORT_FOLDER_PATH"),